ADD BaseNode can get actual view size/pos CHORE Extracted child classes in BaseNode to different files ADD ViewDelegates class which allows Node to Access a views data
