+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.2.7f1
-m_EditorVersionWithRevision: 2021.2.7f1 (6bd9e232123f)
+m_EditorVersion: 2021.3.2f1
+m_EditorVersionWithRevision: 2021.3.2f1 (d6360bedb9a0)
